Change npsAverage from INT to REAL

+++ migrations/migrations.sql
@@ -4085,7 +4085,7 @@ do $$
 
     -- Add npsAverage column to accounts table
     if not exists (select 1 from information_schema.columns where table_name = 'accounts' and column_name = 'npsAverage') then
-      alter table "accounts" add column "npsAverage" INT;
+      alter table "accounts" add column "npsAverage" REAL;
     end if;
 
     -- Add npsUpdatedAt column to accounts table
